+++ src/main/java/topprogersgroup/controller/AdminController.java
@@ -10,13 +10,13 @@ import org.springframework.web.bind.annotation.ModelAttribute;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestMethod;
 import org.springframework.web.servlet.ModelAndView;
-import topprogersgroup.entity.Administrator;
-import topprogersgroup.entity.Employee;
-import topprogersgroup.entity.User;
-import topprogersgroup.entity.UserCreateForm;
+import topprogersgroup.entity.*;
 import topprogersgroup.service.UserService;
 
 import javax.validation.Valid;
+import java.util.Date;
+import java.util.HashMap;
+import java.util.Map;
 
 @Controller
 @RequestMapping(value = "/admin")
@@ -38,13 +38,16 @@ public class AdminController {
         modelAndView.setViewName("admin/user_create");
         modelAndView.addObject("form", new UserCreateForm());
         modelAndView.addObject("admin", new Administrator());
-        modelAndView.addObject("employee", new Employee());
+        Employee e = new Employee();
+        e.setEmploymentDate(new Date());
+        modelAndView.addObject("employee", e);
         return modelAndView;
     }
 
     @PreAuthorize("hasAuthority('ADMIN')")
     @RequestMapping(value = "/create", method = RequestMethod.POST)
-    public String handleUserCreateForm(@Valid @ModelAttribute("form") UserCreateForm form,
+    public String handleUserCreateForm(Model model,
+                                       @Valid @ModelAttribute("form") UserCreateForm form,
                                        BindingResult bindingResult,
                                        @ModelAttribute("admin")Administrator admin,
                                        @ModelAttribute("employee")Employee employee) {
@@ -55,36 +58,63 @@ public class AdminController {
             userService.create(form);
         } catch (DataIntegrityViolationException e) {
             bindingResult.reject("email.exists", "Адрес электронной почты уже существует");
+//            model.a("admin", new Administrator());
+//            Employee e = new Employee();
+//            e.setEmploymentDate(new Date());
+//            modelAndView.addObject("employee", e);
             return "admin/user_create";
         }
-        return "";
+        if(form.getRole().equals(Role.ADMIN)){
+            Administrator admin1= admin;
+        }
+        else if(form.getRole().equals(Role.EMPLOYEE)){
+            Employee employee1 = employee;
+        }
+        return "admin/home";
     }
 
+    @PreAuthorize("hasAuthority('ADMIN')")
+    @RequestMapping(value = {"/vet"}, method = RequestMethod.GET)
+    public String createSVetService(Model model){
+        StateVeterinaryService service = new StateVeterinaryService();
+        model.addAttribute("svService", service);
+        return "admin/statevetservice";
+    }
 
-//    @PreAuthorize("hasAuthority('ADMIN')")
-//    @RequestMapping(value = "/create", method = RequestMethod.GET)
-//    public String getUserCreatePage(Model model) {
-//        model.addAttribute("form", new User());
-//        return "admin/user_create";
-//    }
-//
-//    @PreAuthorize("hasAuthority('ADMIN')")
-//    @RequestMapping(value = "/create", method = RequestMethod.POST)
-//    public String handleUserCreateForm(Model model,
-//                                       @Valid @ModelAttribute("form") User form,
-//                                       BindingResult bindingResult) {
-//        if (bindingResult.hasErrors()) {
-//            model.addAttribute("form", form);
-//            return "admin/user_create";
-//        }
-//        try {
-//            userService.create(form);
-//        } catch (DataIntegrityViolationException e) {
-//            bindingResult.reject("email.exists", "Адрес электронной почты уже существует");
-//            return "admin/user_create";
-//        }
-//        return "";
-//    }
+    @PreAuthorize("hasAuthority('ADMIN')")
+    @RequestMapping(value = {"/vet"}, method = RequestMethod.POST)
+    public String createSVetService(Model model,
+                                    @Valid @ModelAttribute("svService")StateVeterinaryService service,
+                                    BindingResult result){
+        if(result.hasErrors()){
+            model.addAttribute("svService", service);
+            return "admin/statevetservice";
+        }
+        return "forward:admin/home";
+    }
 
+    @PreAuthorize("hasAuthority('ADMIN')")
+    @RequestMapping(value = {"/checkpoint"}, method = RequestMethod.GET)
+    public String createCheckPoint(Model model){
+        CheckPoint checkPoint = new CheckPoint();
+        Map<Integer, Employee> inspector = new HashMap<>();
+        model.addAttribute("checkPoint", checkPoint);
+//        model.addAttribute("inspector", inspector);
+        return "admin/checkpoint";
+    }
+
+    @PreAuthorize("hasAuthority('ADMIN')")
+    @RequestMapping(value = {"/checkpoint"}, method = RequestMethod.POST)
+    public String createCheckPoint(Model model,
+                                   @Valid @ModelAttribute("checkPoint")CheckPoint checkPoint,
+//                                   @ModelAttribute("inspector")Map<Integer, Employee> inspector,
+                                   BindingResult result){
+        if(result.hasErrors()){
+//            model.addAttribute("checkPoint", checkPoint);
+//            model.addAttribute("inspector", inspector);
+            return "admin/checkpoint";
+        }
+        return "forward:admin/home";
+    }
 
 }
