+++ src/main/webapp/admin/home.ftl
@@ -24,8 +24,8 @@
     <#if currentUser?? && currentUser.role == "ADMIN">
         Вы администратор!!
         <li><a href="/admin/create">Добавить пользователя</a></li>
-        <li><a href="/resp/checkpoint">Добавить ПКВП</a></li>
-        <li><a href="/resp/vet">Добавить ГосВет Службу</a></li>
+        <li><a href="/admin/checkpoint">Добавить ПКВП</a></li>
+        <li><a href="/admin/vet">Добавить ГосВет Службу</a></li>
     </#if>
     </ul>
 </nav>
